chore: split `serializer.py` into separate files
