+++ api/views/v1_1.py
@@ -5,9 +5,9 @@ from rest_framework.views import APIView
 
 from anthias_app.models import Asset
 from api.helpers import AssetCreationException, parse_request
+from api.serializers.v1_1 import CreateAssetSerializerV1_1
 from api.serializers import (
     AssetSerializer,
-    CreateAssetSerializerV1_1,
     UpdateAssetSerializer,
 )
 from api.views.v1 import V1_ASSET_REQUEST
